walka ze sprintbach draw -> tzreba przejśc na vektory bo nie ma kontroli z tym co się rysuje

+++ Bomber/BomberManViewModel/Converters/OponentLocationDAOToOponentLocationEntity.cs
@@ -17,8 +17,8 @@ namespace BomberManViewModel.Converters
             Target.Game = Mapper.Map<Game>(Source.Game);
             Target.ID = Source.ID;
             Target.IsAlive = Source.IsAlive;
-            Target.XLocation = Source.XLocation;
-            Target.YLocation = Source.YLocation;
+            Target.XLocation = (int)Source.XLocation;
+            Target.YLocation = (int)Source.YLocation;
             Target.Oponent = Mapper.Map<Oponent>(Source.Oponent);
             return Target;
         }
